Fix missing March VWC in 2012
Discovered that VWC for 2012 is missing the month
of March.  This is a big problem for the analysis
as the analysis uses the sum total VWC for the
spring as the measure of soil moisture.  Thus
2012 is artificially dry.  In this commit
I simply average VWC from April 9th to July 1 each
year.  But this is not ideal.  Potentially need to
use predicted soil moisture for all years from the
SOILWAT model.